Clean up create/edit
